# apps/taptrade-platform/go-platform/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tap Trade API
  version: "1.0.0"
  description: >
    Tap Trade prediction-market gateway API for non-redeemable gameplay points.
    Binary YES/NO outcomes use point-cents subunits on a 0-100 implied
    probability scale; settled winning shares return 100 point-cents per share.
    Points are gameplay-only units with no external value path.

    Coverage: this spec documents the core public, trading, portfolio,
    reward, loyalty, social, leaderboard, bot, auth proxy,
    admin account-review, admin discovery taxonomy, admin market operations,
    and admin market-lifecycle surfaces verified against the live handlers and
    DTOs as of 2026-06. CMS and remaining compatibility surfaces are
    intentionally out of scope here and tracked for documentation in
    IMPROVEMENT_PLAN.md P2-05 / G-04. The previous inherited spec is archived
    outside this launch document.
servers:
  - url: http://localhost:18080
    description: Local development
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly session cookie issued by the auth service. Mutating requests also require the X-CSRF-Token header (double-submit cookie).
    BotApiKey:
      type: http
      scheme: bearer
      description: "Bot/partner API key: Authorization: Bearer tna_<prefix>_<secret>. Scopes: read, trade."
  schemas:
...
    StorePointPack:
      type: object
      description: One purchasable point-pack catalogue entry (point store, STORE_ENABLED only). Purchases credit non-redeemable gameplay points one way; totalPoints is computed from base plus bonus and never stored.
      properties:
        id: { type: string }
        name: { type: string }
        priceUsdCents: { type: integer, format: int64, description: Real point pack purchase price in integer cents. Rendered client-side; never mixed into point fields. }
        basePoints: { type: integer, format: int64, description: Gameplay points included in the point pack purchase. }
        bonusPoints: { type: integer, format: int64, description: Promotional extra gameplay points, shown separately. }
        totalPoints: { type: integer, format: int64, description: Computed base plus bonus gameplay points. }
        displayOrder: { type: integer }
        badge: { type: string, description: Optional label from real pack configuration only (for example Popular or Best Value). }
        introOnly: { type: boolean, description: Pack is served only before the caller's first completed purchase. }
        promoStartsAt: { type: string, format: date-time, nullable: true }
        promoEndsAt: { type: string, format: date-time, nullable: true }
        unit: { type: string, enum: [PTS], description: Gameplay point unit for point-native fields. }
    StorePackListResponse:
      type: object
      properties:
        packs:
          type: array
          items: { $ref: "#/components/schemas/StorePointPack" }
        firstPurchase: { type: boolean, description: True until the caller completes their first point pack purchase. }
        unit: { type: string, enum: [PTS] }
    StorePurchase:
      type: object
      description: One point pack purchase with the immutable server-resolved pack snapshot. Status machine is pending_payment to completed, failed, or canceled; terminal states are final and gameplay points credit exactly once on completion.
      properties:
        id: { type: string, description: Purchase id with the sp_ prefix. }
        userId: { type: string }
        packId: { type: string }
        status: { type: string, enum: [pending_payment, completed, failed, canceled] }
        priceUsdCents: { type: integer, format: int64 }
        basePoints: { type: integer, format: int64 }
        bonusPoints: { type: integer, format: int64 }
        totalPoints: { type: integer, format: int64 }
        provider: { type: string, enum: [demo], description: Checkout provider. The demo provider is a labeled simulation with no real charge; stripe is reserved and refused at boot. }
        providerSessionId: { type: string }
        failureReason: { type: string }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        unit: { type: string, enum: [PTS] }
    StoreCheckoutRequest:
      type: object
      description: The client sends only the pack id and an idempotency key; price and point amounts are resolved server-side from the catalogue row.
      properties:
        packId: { type: string }
        idempotencyKey: { type: string, description: Client idempotency key. A double-submitted checkout returns the same purchase. }
      required: [packId, idempotencyKey]
    StoreCheckoutResponse:
      type: object
      properties:
        purchase: { $ref: "#/components/schemas/StorePurchase" }
        checkout:
          type: object
          properties:
            providerSessionId: { type: string }
            checkoutToken: { type: string, description: Opaque token for the simulated demo checkout screen. }
    StorePurchaseListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/StorePurchase" }
        total: { type: integer }
...
    StoreWebhookEvent:
      type: object
      description: Provider event addressed to a purchase. The body is verified with HMAC-SHA256 over the raw payload (X-Store-Signature header) inside a five-minute timestamp window; amounts never ride the event.
      properties:
        eventId: { type: string }
        eventType: { type: string }
        purchaseId: { type: string }
        status: { type: string, enum: [completed, failed, canceled, pending] }
        timestamp: { type: string, description: Unix seconds or RFC3339. }
      required: [purchaseId, status, timestamp]
    StoreWebhookAck:
      type: object
      properties:
        purchaseId: { type: string }
        status: { type: string }
        alreadyFulfilled: { type: boolean }
...
paths:
...
  /api/v1/store/packs:
    get:
      summary: List the purchasable point-pack catalogue for the caller
      description: Point store surface, registered only when STORE_ENABLED is true. Serves active packs inside their promo windows with intro-only eligibility applied, plus the caller's first-purchase flag. Point pack purchases credit non-redeemable gameplay points one way.
      security: [{ SessionCookie: [] }]
      responses:
        "200":
          description: Point pack catalogue with eligibility
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorePackListResponse" }
        "403": { description: Authentication required }
  /api/v1/store/checkout:
    post:
      summary: Create a point pack checkout for the caller
      description: Resolves price and point amounts server-side from the pack row (the client supplies only packId plus an idempotency key), snapshots them onto a pending purchase, and registers a provider checkout session. A double-submitted idempotency key returns the same purchase with status 200.
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoreCheckoutRequest" }
      responses:
        "201":
          description: New pending purchase with provider checkout session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoreCheckoutResponse" }
        "200":
          description: Idempotent replay returning the existing purchase
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoreCheckoutResponse" }
        "403": { description: Authentication required or pack not currently purchasable }
        "404": { description: Point pack not found }
  /api/v1/store/purchases:
    get:
      summary: List the caller's point pack purchase history
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: Purchase history for the session user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorePurchaseListResponse" }
        "403": { description: Authentication required }
...
  /api/v1/store/webhook:
    post:
      summary: Receive a signature-verified provider event
      description: Public and CSRF-exempt only while STORE_ENABLED is true. The handler verifies an HMAC-SHA256 signature over the raw body (X-Store-Signature) within a five-minute timestamp window before processing; replays are idempotent no-ops backed by the purchase state machine and point ledger keys.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoreWebhookEvent" }
      responses:
        "200":
          description: Event applied (or replayed idempotently)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoreWebhookAck" }
        "401": { description: Invalid webhook signature }
        "404": { description: Purchase not found }
        "503": { description: Webhook secret not configured }
...